Add basic flask server.
The server now serves two simple pages (the map and the horus) which
have layouts similar to the original diogenet application.

In a similar change, webpack on the client now produces two bundles
instead of one, each bundle corresponds to one of the pages (so
`map.bundle.js` and `horus.bundle.js`). Bootstrap was added for now
to the `static/` directory of the flask server, but it should be
moved to the client with a custom build to support theming and to
be able to handle bootstrap's javascript as a module.